requirements.txt
annotated-types==0.7.0
anyio==4.10.0
AnyQt==0.2.1
asttokens==3.0.0
attrs==25.3.0
baycomp==1.0.3
blinker==1.9.0
Bottleneck==1.5.0
cattrs==25.1.1
certifi==2025.8.3
chardet==5.2.0
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
comm==0.2.3
commonmark==0.9.1
contourpy==1.3.3
cycler==0.12.1
debugpy==1.8.16
decorator==5.2.1
dictdiffer==0.9.0
docutils==0.22
et_xmlfile==2.0.0
executing==2.2.0
fastapi==0.116.1
Flask==3.1.1
fonttools==4.59.0
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
ipykernel==6.30.1
ipython==9.4.0
ipython_pygments_lexers==1.1.1
itsdangerous==2.2.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.2.1
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.1
jupyter_client==8.6.3
jupyter_core==5.8.1
keyring==25.6.0
keyrings.alt==5.0.2
kiwisolver==1.4.9
MarkupSafe==3.0.2
matplotlib==3.10.5
matplotlib-inline==0.1.7
more-itertools==10.7.0
nest-asyncio==1.6.0
networkx==3.5
numpy==2.3.1
openpyxl==3.1.5
openTSNE==1.0.2
orange-canvas-core==0.2.6
orange-widget-base==4.26.0
Orange3==3.39.0
packaging==25.0
pandas==2.3.1
parso==0.8.4
pillow==11.3.0
platformdirs==4.3.8
prompt_toolkit==3.0.51
psutil==7.0.0
pure_eval==0.2.3
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
pyparsing==3.2.3
pyqtgraph==0.13.7
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-louvain==0.16
pytz==2025.2
pywin32==311
pywin32-ctypes==0.2.3
PyYAML==6.0.2
pyzmq==27.0.1
qasync==0.27.1
qtconsole==5.6.1
QtPy==2.4.3
requests==2.32.4
requests-cache==1.2.1
scikit-learn==1.7.0
scipy==1.16.0
serverfiles==0.3.1
six==1.17.0
sniffio==1.3.1
stack-data==0.6.3
starlette==0.47.3
threadpoolctl==3.6.0
tornado==6.5.2
traitlets==5.14.3
truststore==0.10.4
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
url-normalize==2.2.1
urllib3==2.5.0
uvicorn==0.35.0
watchfiles==1.1.0
wcwidth==0.2.13
websockets==15.0.1
Werkzeug==3.1.3
xgboost==2.0.3
xlrd==2.0.2
xlsxwriter==3.2.5
